still WIP, mostly finished - will complete final changes tomorrow + proper commit message (+1 squashed commits)
Squashed commits:

[a9f183b] WORKING - only problem is json arrays, deleted keys and updates which change object. This is semantic problems, so need to talk to Rob about this.
Added temporary tests in conflict, will move to anothe test after.

Added :
updateHistory - creates an object blame json file. Stores this on a branch off relevant commits.
updateHistory method basically follows the commit path and creates a history branch on the most recent commit, allowing blame etc on all configuration changes. It uses any history branches it can find on this path to reduce the time taken, so previous updates to the history are used, saving lots of time.
It also saves copies of the history on any commits which are used by multiple branches, saving time for getting history on these branches (to add (takes like 1 line)) [ note that the path taken is very important here, so the commit must contain path taken, and this must be compatible with the other branch updating history, or we must start the history again. ] (+2 squashed commit)

Squashed commit:

[838e632] WIP ~ Object history graph using history branches

[2c2caf0] WIP
- Adding a blame object graph through creating history branches throughout the git repository. Implemented a number of memoization methods to stop the entire commit tree being checked out.
 - Lots to add, with the main feature missing being the Json object comparisons and creation. Most of framework around this added however.